Added more java code
