+++ app/src/main/res/layout/tractor_screen.xml
@@ -131,15 +131,17 @@
             </FrameLayout>
 
             <ImageButton
-                android:id="@+id/imageButton"
+                android:id="@+id/image_back_button"
                 android:layout_width="wrap_content"
+                android:clickable="true"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="4dp"
                 android:background="@android:color/transparent"
                 android:contentDescription="@string/Back_button"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:srcCompat="@drawable/group_15" />
+                app:srcCompat="@drawable/group_15"
+                android:focusable="true" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
     </ScrollView>
